fix(faiss_db): move get_relevant_sample_asks from assistant to faiss_db module
